把提取数据部分写好了

+++ auto_updata.h
@@ -1,6 +1,6 @@
 #ifndef _AUTO_UPDATA_H_
 #define _AUTO_UPDATA_H_
 
-int updata_main(int );
+int updata_main(double );
 
 #endif
